Move test project into build infrastructure
