--- src/Grpc/Directory.Build.props
@@ -1,2 +1,0 @@
-<Project>
-</Project>
