--- src/Grpc/Directory.Build.targets
@@ -1,2 +1,0 @@
-<Project>
-</Project>
