+++ src/Grpc/test/InteropTests/Helpers/ClientProcess.cs
@@ -1,20 +1,21 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Diagnostics;
-using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Internal;
 using Xunit.Abstractions;
 
-namespace InteropTests.Infrastructure
+namespace InteropTests.Helpers
 {
-    public class WebServerProcess : IDisposable
+    public class ClientProcess : IDisposable
     {
         private readonly Process _process;
         private readonly ProcessEx _processEx;
         private readonly TaskCompletionSource<object> _startTcs;
-        private readonly StringBuilder _output;
 
-        public WebServerProcess(string path, ITestOutputHelper output)
+        public ClientProcess(ITestOutputHelper output, string path, int port, string testCase)
         {
             _process = new Process();
             _process.StartInfo = new ProcessStartInfo
@@ -22,7 +23,7 @@ namespace InteropTests.Infrastructure
                 RedirectStandardOutput = true,
                 RedirectStandardError = true,
                 FileName = "dotnet.exe",
-                Arguments = @$"run -p {path}"
+                Arguments = @$"run -p {path} --use_tls false --server_port {port} --client_type httpclient --test_case {testCase}"
             };
             _process.EnableRaisingEvents = true;
             _process.OutputDataReceived += Process_OutputDataReceived;
@@ -31,26 +32,21 @@ namespace InteropTests.Infrastructure
             _processEx = new ProcessEx(output, _process);
 
             _startTcs = new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously);
-            _output = new StringBuilder();
         }
 
         public Task WaitForReady()
         {
-            if (_processEx.HasExited)
-            {
-                return Task.FromException(new InvalidOperationException("Server is not running."));
-            }
-
             return _startTcs.Task;
         }
+
+        public int ExitCode => _process.ExitCode;
+        public Task Exited => _processEx.Exited;
 
         private void Process_OutputDataReceived(object sender, DataReceivedEventArgs e)
         {
             var data = e.Data;
             if (data != null)
             {
-                _output.AppendLine(data);
-
                 if (data.Contains("Application started."))
                 {
                     _startTcs.TrySetResult(null);
