--- src/Grpc/test/InteropTests/Infrastructure/TaskExtensions.cs
@@ -1,66 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace InteropTests.Infrastructure
-{
-    public static class TaskExtensions
-    {
-        public static async Task<T> TimeoutAfter<T>(this Task<T> task, TimeSpan timeout,
-            [CallerFilePath] string filePath = null,
-            [CallerLineNumber] int lineNumber = default)
-        {
-            // Don't create a timer if the task is already completed
-            // or the debugger is attached
-            if (task.IsCompleted || Debugger.IsAttached)
-            {
-                return await task;
-            }
-
-            var cts = new CancellationTokenSource();
-            if (task == await Task.WhenAny(task, Task.Delay(timeout, cts.Token)))
-            {
-                cts.Cancel();
-                return await task;
-            }
-            else
-            {
-                throw new TimeoutException(CreateMessage(timeout, filePath, lineNumber));
-            }
-        }
-
-        public static async Task TimeoutAfter(this Task task, TimeSpan timeout,
-            [CallerFilePath] string filePath = null,
-            [CallerLineNumber] int lineNumber = default)
-        {
-            // Don't create a timer if the task is already completed
-            // or the debugger is attached
-            if (task.IsCompleted || Debugger.IsAttached)
-            {
-                await task;
-                return;
-            }
-
-            var cts = new CancellationTokenSource();
-            if (task == await Task.WhenAny(task, Task.Delay(timeout, cts.Token)))
-            {
-                cts.Cancel();
-                await task;
-            }
-            else
-            {
-                throw new TimeoutException(CreateMessage(timeout, filePath, lineNumber));
-            }
-        }
-
-        private static string CreateMessage(TimeSpan timeout, string filePath, int lineNumber)
-            => string.IsNullOrEmpty(filePath)
-            ? $"The operation timed out after reaching the limit of {timeout.TotalMilliseconds}ms."
-            : $"The operation at {filePath}:{lineNumber} timed out after reaching the limit of {timeout.TotalMilliseconds}ms.";
-    }
-}
