+++ src/Grpc/test/testassets/Directory.Build.props
@@ -1,0 +1,3 @@
+<Project>
+  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory)..\, Directory.Build.props))\eng\Versions.props" />
+</Project>
