+++ src/Grpc/test/testassets/Directory.Build.targets
@@ -1,0 +1,1 @@
+<Project />
