+++ src/Grpc/test/testassets/InteropTestsWebsite/InteropTestsWebsite.csproj
@@ -13,7 +13,7 @@
     <Protobuf Include="..\Proto\grpc\testing\empty.proto" GrpcServices="None" Link="Protos\empty.proto" />
     <Protobuf Include="..\Proto\grpc\testing\messages.proto" GrpcServices="None" Link="Protos\messages.proto" />
 
-    <PackageReference Include="Grpc.AspNetCore" Version="2.25.0" />
+    <PackageReference Include="Grpc.AspNetCore" Version="$(GrpcAspNetCorePackageVersion)" />
   </ItemGroup>
 
 </Project>
